requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0          # PostgreSQL async driver
aiosqlite>=0.20.0        # SQLite async driver (tests)
alembic>=1.13.0

# Auth / security
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
passlib>=1.7.4           # kept for compatibility (security.py uses bcrypt directly)

# Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0

# File I/O
aiofiles>=23.0.0
python-multipart>=0.0.9   # required by FastAPI for form data / file uploads

# Templating
jinja2>=3.1.0

# Markup rendering
mistune>=3.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
